+++ src/main/java/com/flab/oasis/mapper/feed/FeedMapper.java
@@ -1,0 +1,9 @@
+package com.flab.oasis.mapper.feed;
+
+import com.flab.oasis.model.TestModel;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface FeedMapper {
+    public TestModel dbConnectionTest();
+}
