+++ src/main/java/com/flab/oasis/service/HomeService.java
@@ -1,7 +1,7 @@
 package com.flab.oasis.service;
 
 import com.flab.oasis.constant.BookCategory;
-import com.flab.oasis.mapper.UserCategoryMapper;
+import com.flab.oasis.mapper.user.UserCategoryMapper;
 import com.flab.oasis.model.BookSuggestion;
 import com.flab.oasis.model.BookSuggestionRequest;
 import com.flab.oasis.model.UserCategory;
