+++ src/main/resources/mapper/book/TestMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.flab.oasis.mapper.TestMapper">
+<mapper namespace="com.flab.oasis.mapper.book.TestMapper">
 
     <select id="dbConnectionTest" resultType="com.flab.oasis.model.TestModel">
         SELECT 10 DATA
