+++ src/main/resources/mapper/user/UserAuthMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.flab.oasis.mapper.UserAuthMapper">
+<mapper namespace="com.flab.oasis.mapper.user.UserAuthMapper">
 
     <select id="getUserAuthByUid" parameterType="java.lang.String" resultType="com.flab.oasis.model.UserAuth">
         SELECT UID,
