+++ src/main/resources/mapper/user/UserCategoryMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.flab.oasis.mapper.UserCategoryMapper">
+<mapper namespace="com.flab.oasis.mapper.user.UserCategoryMapper">
 
     <select id="findUserCategoryByUid" parameterType="java.lang.String" resultType="com.flab.oasis.model.UserCategory">
         SELECT UID, BOOK_CATEGORY AS bookCategory
